Move element bean definition builders to processing package
