+++ core-processor/src/main/java/io/micronaut/aop/writer/ProxyingBeanDefinitionWriter.java
@@ -22,9 +22,9 @@ import io.micronaut.core.annotation.AnnotationMetadata;
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.value.OptionalValues;
-import io.micronaut.inject.ElementBeanDefinitionBuilder;
-import io.micronaut.inject.ElementProxyBuilder;
-import io.micronaut.inject.OutputObjectDef;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilder;
+import io.micronaut.inject.processing.definition.ElementProxyBuilder;
+import io.micronaut.inject.processing.definition.OutputObjectDef;
 import io.micronaut.inject.ProxyBeanDefinition;
 import io.micronaut.inject.annotation.MutableAnnotationMetadata;
 import io.micronaut.inject.ast.ClassElement;
