+++ core-processor/src/main/java/io/micronaut/inject/processing/ConfigurationReaderBeanElementCreator.java
@@ -26,8 +26,8 @@ import io.micronaut.context.beans.definition.MethodDefinition;
 import io.micronaut.core.annotation.AnnotationMetadata;
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
-import io.micronaut.inject.ElementBeanDefinitionBuilder;
-import io.micronaut.inject.ElementBeanDefinitionBuilderFactory;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilder;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilderFactory;
 import io.micronaut.inject.annotation.AnnotationMetadataHierarchy;
 import io.micronaut.inject.annotation.MutableAnnotationMetadata;
 import io.micronaut.inject.ast.ClassElement;
