+++ core-processor/src/main/java/io/micronaut/inject/processing/definition/ElementProxyBuilder.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2017-2026 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.inject.processing.definition;
+
+import io.micronaut.context.beans.definition.ProxyBuilder;
+import io.micronaut.inject.ast.ClassElement;
+import io.micronaut.inject.ast.MethodElement;
+import io.micronaut.inject.writer.OriginatingElements;
+
+import java.util.List;
+
+/**
+ * Proxy builder that operates on {@link ClassElement} definitions.
+ *
+ * @param <R> The builder result type
+ * @author Denis Stepanov
+ * @since 5.1.0
+ */
+public interface ElementProxyBuilder<R> extends ProxyBuilder<ClassElement, MethodElement, List<R>>, OriginatingElements {
+
+    /**
+     * @return The underlying bean definition builder used to materialize the proxy
+     */
+    ElementBeanDefinitionBuilder<R> beanDefinitionBuilder();
+
+}
