+++ core-processor/src/main/java/io/micronaut/inject/writer/AbstractBeanDefinitionBuilder.java
@@ -26,9 +26,9 @@ import io.micronaut.core.annotation.AnnotationValueBuilder;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.util.ArrayUtils;
 import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.inject.ElementBeanDefinitionBuilder;
-import io.micronaut.inject.ElementBeanDefinitionBuilderFactory;
-import io.micronaut.inject.ElementProxyBuilder;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilder;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilderFactory;
+import io.micronaut.inject.processing.definition.ElementProxyBuilder;
 import io.micronaut.inject.annotation.AnnotationMetadataHierarchy;
 import io.micronaut.inject.annotation.MutableAnnotationMetadata;
 import io.micronaut.inject.ast.ClassElement;
