+++ inject-groovy/src/main/groovy/io/micronaut/ast/groovy/visitor/GroovyBeanDefinitionBuilder.java
@@ -21,8 +21,8 @@ import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.AnnotationValueBuilder;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.util.ArgumentUtils;
-import io.micronaut.inject.ElementBeanDefinitionBuilder;
-import io.micronaut.inject.ElementBeanDefinitionBuilderFactory;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilder;
+import io.micronaut.inject.processing.definition.ElementBeanDefinitionBuilderFactory;
 import io.micronaut.inject.annotation.AnnotationMetadataHierarchy;
 import io.micronaut.inject.ast.ClassElement;
 import io.micronaut.inject.ast.Element;
